+++ test/core-model-test/src/main/java/com/ainrif/apiator/test/model/core/Dto07_JavaDocField.java
@@ -18,7 +18,6 @@ package com.ainrif.apiator.test.model.core;
 
 public class Dto07_JavaDocField {
     /**
-     * todo #not_supported
      * Field-level javadoc
      */
     public Object objectField;